Moved blog models into their own app
