+++ tbx/blog/templates/blog/blog_index_page.html
@@ -37,7 +37,7 @@
         </div>
 
         <div class="container">
-            {% include "torchbox/includes/blog_listing.html" %}
+            {% include "blog/includes/blog_listing.html" %}
         </div>
 
     </section>
