+++ tbx/core/urls.py
@@ -1,7 +1,7 @@
 from django.conf.urls import url
 
+from tbx.blog.feeds import BlogFeed, PlanetDrupalFeed
 from tbx.core import views
-from tbx.core.feeds import BlogFeed, PlanetDrupalFeed
 
 urlpatterns = [
     url(r'^blog/feed/$', BlogFeed(), name='blog_feed'),
